gracefully handle peer/optional dependencies not being installed
