+++ base/index.js
@@ -1,7 +1,11 @@
+const { pluginConfigs } = require('../utils/plugins');
+
 module.exports = {
   extends: [
     require.resolve('./eslint.js'),
-    require.resolve('./jest.js'),
-    require.resolve('./import.js')
+    ...pluginConfigs([
+      'jest',
+      'import',
+    ])
   ],
 };
